EZP-30949: changed tag name and namespace
